Relocate dbConfig file; return obj rather than fn
Resolves #96
