+++ careen.js
@@ -1,15 +1,17 @@
 const path = require('path');
-const dbConfig = require('./server/api/util/db-config');
+const dbConfig = require('./config/db-config');
 
-var config = dbConfig();
+var config;
 
 // Unfortunately, careen has a different signature than node-postgres.
 // The dbConfig method returns correct version for node-postgres, so we
 // need to manipulate it here.
-if (typeof config === 'string') {
+if (typeof dbConfig === 'string') {
   config = {
-    url: config
+    url: dbConfig
   };
+} else {
+  config = dbConfig;
 }
 
 module.exports = {
